requirements.txt
pandas
numpy
matplotlib
mplfinance
ta
yfinance
requests
python-telegram-bot
